Move IPGetOffset and IPAddOffset to iphelpers package
