+++ pkg/allocate/allocate_test.go
@@ -2,7 +2,6 @@ package allocate
 
 import (
 	"fmt"
-	"math"
 	"math/big"
 	"net"
 	"testing"
@@ -17,152 +16,6 @@ func TestAllocate(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "cmd")
 }
-
-var _ = Describe("Allocation utility functions", func() {
-	/*
-		func byteSliceAdd(ar1, ar2 []byte) ([]byte, error) {
-		func byteSliceSub(ar1, ar2 []byte) ([]byte, error) {
-		func ipAddrToUint64(ip net.IP) uint64 {
-		func ipAddrFromUint64(num uint64) net.IP {
-		func IPAddOffset(ip net.IP, offset uint64) net.IP {
-		func IPGetOffset(ip1, ip2 net.IP) uint64 {
-	*/
-	It("tests byteSliceAdd normal case", func() {
-		b1 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1}
-		b2 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 10}
-		bSum, err := byteSliceAdd(b1, b2)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(bSum).To(Equal([]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 11, 11}))
-	})
-
-	It("tests byteSliceAdd carry case", func() {
-		b1 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255}
-		b2 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1}
-		bSum, err := byteSliceAdd(b1, b2)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(bSum).To(Equal([]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0}))
-	})
-
-	It("tests byteSliceSub normal case", func() {
-		b1 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 10}
-		b2 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1}
-		bSum, err := byteSliceSub(b1, b2)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(bSum).To(Equal([]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 9, 9}))
-	})
-
-	It("tests byteSliceSub carry case", func() {
-		b1 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0}
-		b2 := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1}
-		bSum, err := byteSliceSub(b1, b2)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(bSum).To(Equal([]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255}))
-	})
-
-	It("can convert ipAddrToUint64", func() {
-		b := []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 255, 255}
-		bNum := ipAddrToUint64(net.IP(b))
-		Expect(bNum).To(Equal(uint64(0x1ffff)))
-	})
-
-	It("can convert ipAddrFromUint64", func() {
-		uintNum := uint64(0x1ffff)
-		ip := net.IP([]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 255, 255})
-		bNum := ipAddrFromUint64(uintNum)
-		Expect(bNum).To(Equal(ip))
-	})
-
-	It("confirms the IPAddOffset normal case", func() {
-		ip1 := net.ParseIP("192.168.2.23")
-		ip2 := IPAddOffset(ip1, uint64(1))
-		Expect(ip2).To(Equal(net.ParseIP("192.168.2.24")))
-
-		ip3 := net.ParseIP("ff02::1")
-		ip4 := IPAddOffset(ip3, uint64(1))
-		Expect(ip4).To(Equal(net.ParseIP("ff02::2")))
-	})
-
-	It("confirms the IPAddOffset carry case", func() {
-		ip1 := net.ParseIP("192.168.2.255")
-		ip2 := IPAddOffset(ip1, uint64(1))
-		Expect(ip2).To(Equal(net.ParseIP("192.168.3.0")))
-
-		ip3 := net.ParseIP("192.168.255.255")
-		ip4 := IPAddOffset(ip3, uint64(1))
-		Expect(ip4).To(Equal(net.ParseIP("192.169.0.0")))
-
-		ip5 := net.ParseIP("ff02::ff")
-		ip6 := IPAddOffset(ip5, uint64(1))
-		Expect(ip6).To(Equal(net.ParseIP("ff02::0:100")))
-
-		ip7 := net.ParseIP("ff02::ffff")
-		ip8 := IPAddOffset(ip7, uint64(1))
-		Expect(ip8).To(Equal(net.ParseIP("ff02::1:0")))
-
-		ip9 := net.ParseIP("ff02::ffff:ffff")
-		ip10 := IPAddOffset(ip9, uint64(1))
-		Expect(ip10).To(Equal(net.ParseIP("ff02::1:0:0")))
-	})
-
-	It("confirms the IPAddOffset error case", func() {
-		ip1 := net.ParseIP("192.168.2.255")
-		ip2 := IPAddOffset(ip1, uint64(math.MaxUint32+1))
-		Expect(ip2).To(BeNil())
-
-		ip3 := net.ParseIP("ff02::1")
-		ip4 := IPAddOffset(ip3, uint64(math.MaxUint32+1))
-		Expect(ip4).NotTo(BeNil())
-	})
-
-	It("confirms the IPGetOffset normal case", func() {
-		ip1 := net.ParseIP("192.168.2.255")
-		ip2 := net.ParseIP("192.168.2.1")
-		offset1 := IPGetOffset(ip1, ip2)
-		Expect(offset1).To(Equal(uint64(254)))
-
-		ip3 := net.ParseIP("ff02::ff")
-		ip4 := net.ParseIP("ff02::1")
-		offset2 := IPGetOffset(ip3, ip4)
-		Expect(offset2).To(Equal(uint64(254)))
-	})
-
-	It("confirms the IPGetOffset carry case", func() {
-		ip1 := net.ParseIP("192.168.3.0")
-		ip2 := net.ParseIP("192.168.2.1")
-		offset1 := IPGetOffset(ip1, ip2)
-		Expect(offset1).To(Equal(uint64(255)))
-
-		ip3 := net.ParseIP("ff02::100")
-		ip4 := net.ParseIP("ff02::1")
-		offset2 := IPGetOffset(ip3, ip4)
-		Expect(offset2).To(Equal(uint64(255)))
-
-		ip5 := net.ParseIP("ff02::1:0")
-		ip6 := net.ParseIP("ff02::1")
-		offset3 := IPGetOffset(ip5, ip6)
-		Expect(offset3).To(Equal(uint64(0xffff)))
-	})
-
-	It("confirms the IPGetOffset error case", func() {
-		// cannot get offset from v4/v6
-		ip1 := net.ParseIP("192.168.3.0")
-		ip2 := net.ParseIP("ff02::1")
-		offset1 := IPGetOffset(ip1, ip2)
-		Expect(offset1).To(Equal(uint64(0)))
-
-		// cannot get offset from v6/v4
-		ip3 := net.ParseIP("ff02::1")
-		ip4 := net.ParseIP("192.168.3.0")
-		offset2 := IPGetOffset(ip3, ip4)
-		Expect(offset2).To(Equal(uint64(0)))
-
-		// cannot get offset between different length array (ar[4] and ar[16])
-		ip5 := net.ParseIP("192.168.4.0")
-		ip6 := net.ParseIP("192.168.3.0")
-		offset3 := IPGetOffset(ip5.To4(), ip6)
-		Expect(offset3).To(Equal(uint64(0)))
-	})
-})
 
 var _ = Describe("Allocation operations", func() {
 	It("creates an IPv4 range properly for 30 bits network address", func() {
